server/requirements.txt
flask
requests
flask_cors
mysql-connector-python